requirements.txt
PyPDF2==3.0.1
langchain==0.0.267
openai==0.27.4
streamlit==1.26.0
python-dotenv==1.0.0
tiktoken==0.3.3
faiss-cpu==1.7.4
